requirements.txt
streamlit==1.32.0
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
easyocr
